refactor(errors): move PreconditionError base to shared/ (AD-2), re-export from core
